refactor(downloads): extract materialization ownership
