+++ docs/generated-testing.md
@@ -41,7 +41,7 @@ Hunting — Generated-First".
 | `tests/test_multidisc_manifest_generated.py` | `try_multi_enqueue` with the REAL `check_for_match` matcher over generated multi-disc worlds | **the #550 coverage law**: an accepted multi-disc grab has no duplicate transfer identities, one distinct source folder per disc, unique coverage == track count. This harness **found the live #550 defect-#1 bug** (partial-disc manifests) before the production MANIFEST-TRACE window captured it |
 | `tests/test_import_manifest_generated.py` | `build_active_download_state` ⇄ `lib/download_reconstruction.py::reconstruct_grab_list_entry` + `check_audio_manifest` | manifest round-trip through persisted state never shrinks or grows (exact key/field fidelity); on-disk check oracle — pass iff disk audio set == tracked manifest set, extra/missing reported exactly |
 | `tests/test_dispatch_outcomes_generated.py` | `dispatch_import_core` + every rejection writer (`_reject_import_from_evidence_decision`, `_record_rejection_and_maybe_requeue`, `DatabaseSource.reject_and_requeue`, `_reject_request_auto_import`, and the atomic abandoned-import writer) | the auditability law (every outcome writes a download_log row); full routing oracle vs `dispatch_action` across the decision universe + no-JSON crash path; U11 self-heal override for ALL 5 preimport facts (incl. `mixed_source`, missing from the hand-written table); every rejection writer preserves the `ValidationResult` distance/scenario projection exactly, including NULL and measured 0.0 |
-| `tests/test_materialize_generated.py` | `lib/processing_paths.py::attempt_fingerprint` + `canonical_folder_for_row` / `canonical_processing_path`, and `lib/download_processing.py::_materialize_processing_dir` | #550 phase 2 follow-up (PR #560 shipped deterministic-only): `attempt_fingerprint` is permutation-invariant, deterministic, distinguishes different `(username, filename)` sets, and the empty set is a stable defined digest; the fingerprint suffix is present iff the fingerprint is non-empty and the resulting basename stays ≤255 bytes even under adversarial generated unicode artist/title (r2 truncation guard); the materialize isolation law — two attempts for the same artist/title/year, driven through the real `_materialize_processing_dir` against a real tempdir, never blend files (attempt B's folder holds exactly B's manifest, A's folder is untouched) and identical manifests resolve to the same folder (resume stability) |
+| `tests/test_materialize_generated.py` | `lib/processing_paths.py::attempt_fingerprint` + `canonical_folder_for_row` / `canonical_processing_path`, and `lib/download_materialization.py::_materialize_processing_dir` | #550 phase 2 follow-up (PR #560 shipped deterministic-only): `attempt_fingerprint` is permutation-invariant, deterministic, distinguishes different `(username, filename)` sets, and the empty set is a stable defined digest; the fingerprint suffix is present iff the fingerprint is non-empty and the resulting basename stays ≤255 bytes even under adversarial generated unicode artist/title (r2 truncation guard); the materialize isolation law — two attempts for the same artist/title/year, driven through the real `_materialize_processing_dir` against a real tempdir, never blend files (attempt B's folder holds exactly B's manifest, A's folder is untouched) and identical manifests resolve to the same folder (resume stability) |
 | `tests/test_quarantine_triage_generated.py` | `lib.quarantine_triage_service.list_unreferenced_quarantine_folders` | quarantine lifecycle law across request statuses: the deterministic result is exactly the immediate real album folders without a default-visible relative/absolute/descendant Wrong Matches reference; replaced frozen-audit rows and external references do not claim local folders, and the code-owned `bad_files` / `untracked_audio` roots are never surfaced or recursively traversed |
 | `tests/test_wrong_match_policy_generated.py` | `lib.wrong_match_policy`, `lib.wrong_matches.wrong_match_row_is_visible`, and `FakePipelineDB.get_wrong_matches` | Wrong Matches remains a candidate/pressing review queue: all five folder/audio-integrity fact rejects plus `spectral_reject` stay excluded across generated scenario/status/history-view worlds, while arbitrary other or NULL scenarios remain visible unless the request is replaced in the default view; independently pinned oracles and known-bad checkers prevent the taxonomy, row predicate, and fake from drifting together |
 | `tests/test_convergence_runner_generated.py` | `lib/convergence.py::run_convergence_steps` | every registered convergence step is attempted exactly once in declared order even when any arbitrary subset raises; Phase 0 order and the end-of-cycle harvest-before-purge constraint are pinned as registry data rather than source inspection; import failures are isolated like call failures |
